requirements.txt
gitdb==4.0.10
GitPython==3.1.29
python-dotenv==0.21.0
smmap==5.0.0
